requirements.txt
python-dotenv
Pillow
opencv-python
imageio
moviepy
numpy
pydub
pytube
requests
tqdm
youtube-search-python
youtube-dl 
google-auth 
google-auth-oauthlib 
google-auth-httplib2
google-api-python-client